+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Decouple `AnnotationsUtil` from `com.google.auto:auto-common`. This involved importing and reimplementing `AnnotationsUtil.getAnnotationValuesWithDefaults()` from `auto-common`. The original `auto-common` 
 * Expose `AbstractStandardProcessor.reportError(...)` method as protected access so subclasses can invoke.
 * Add `AbstractStandardProcessor.emitTypeSpec(...)` helper method as it appeared in all downstream classes. In the future it will also make it possible for the toolkit to track processing statistics.
+* Decouple `AbstractStandardProcessor` from `com.google.auto:auto-common`.
+* Remove unused dependencies `com.google.auto:auto-common` and `com.google.guava:guava`.
 
 ### [v0.06](https://github.com/realityforge/proton/tree/v0.06) (2019-12-25) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.05...v0.06)